fix: mobile header slide animation
